Back_End/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
python-multipart
cryptography
bcrypt
pypdfium2
pdfplumber
pandas
numpy
langchain
langchain-openai
langchain-chroma
langchain-text-splitters
langchain-community
langchain-core
langgraph
langgraph-checkpoint-sqlite
pydantic
openai
chromadb
pyjwt
